Add/Update/Fix/Run RPC server and client tests (#214)
Makes the existing tests run on build and actually work.
Adds test for Greengrass client V2.

This change very crucially adds a call to builder.excludeFieldsWithoutExposeAnnotation() when constructing the GSON instance. This tells GSON to only serialize and deserialize fields which are marked as @Exposed. This is necessary for JDK 9+ specifically to serialize errors because without it, GSON tries to serialize private fields from Throwable which Java prevents.

Fixes Greengrass client V2 which didn't run lambda-based callbacks in the executor.
Makes the socket domain configurable in the v2 client builder so that we can use IP instead of domain sockets.