First pass at moving command line parsing to single file
Added a new Java class for command line parsing based on the C++ command
line parsing file. Adjusted all samples to use the new file so the samples
are smaller, have less duplicate code, and are easier to read.

Also adjusted the tests to use the correct command line names.