refactor(greengrass): add with-ers and remove the return from setters which broke compatibility (#223)
In release 1.6.0, we updated the Greengrass data model setters so that they return `this` so that calls could be chained. Unfortunately this broke binary compatibility which is necessary for Greengrass to maintain backward compatibility. This change should be released as version 1.7.0; it is compatible with versions <1.6.0. Users who are using >=1.6.0, will be able to upgrade to this version if they are not using the return value from the setters. They can also simply swap the calls to setX with a call to withX.